Server context file location updated; TLS protocol added

+++ zRageMapDownloader.UI/App.xaml.cs
@@ -4,6 +4,7 @@ using System.Configuration;
 using System.Data;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using System.Windows;
 using zRageMapDownloader.Core;
@@ -17,13 +18,12 @@ namespace zRageMapDownloader
     {
         public App()
         {
-            try
+            ServicePointManager.Expect100Continue = true;
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
+
+            if (Directory.Exists(MapManager.MainTempFolder))
             {
                 Directory.Delete(MapManager.MainTempFolder, true);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
             }
         }
     }
